+++ packages/react-native/src/sync/stream/ReactNativeStreamingSyncImplementation.ts
@@ -2,9 +2,9 @@ import {
   AbstractStreamingSyncImplementation,
   AbstractStreamingSyncImplementationOptions,
   LockOptions,
-  LockType
+  LockType,
+  Mutex
 } from '@powersync/common';
-import { Mutex } from 'async-mutex';
 
 /**
  * Global locks which prevent multiple instances from syncing
@@ -45,11 +45,7 @@ export class ReactNativeStreamingSyncImplementation extends AbstractStreamingSyn
       throw new Error(`Lock type ${lockOptions.type} not found`);
     }
     return lock.runExclusive(async () => {
-      if (lockOptions.signal?.aborted) {
-        throw new Error('Aborted');
-      }
-
       return lockOptions.callback();
-    });
+    }, lockOptions.signal);
   }
 }