+++ packages/powersync-op-sqlite/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @powersync/op-sqlite
 
+## 0.9.4
+
+### Patch Changes
+
+- 8f8ef1c: Remove `async-mutex` dependency in favor of internal implementation.
+- Updated dependencies [eac163a]
+- Updated dependencies [8f8ef1c]
+  - @powersync/common@1.50.0
+
 ## 0.9.3
 
 ### Patch Changes